refactor : s3 config 관련 파일 infrastructure -> common/config 디렉토리로 변경 (#559)
