+++ src/main/java/com/clubber/ClubberServer/domain/image/controller/ImageController.java
@@ -3,12 +3,11 @@ package com.clubber.ClubberServer.domain.image.controller;
 import com.clubber.ClubberServer.domain.image.dto.CreateImagePresignedUrlResponse;
 import com.clubber.ClubberServer.domain.image.dto.CreateRecruitsImagePresigneUrlRequest;
 import com.clubber.ClubberServer.domain.image.service.S3UploadPresignedService;
-import com.clubber.ClubberServer.global.infrastructure.s3.ImageFileExtension;
+import com.clubber.ClubberServer.global.config.s3.ImageFileExtension;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
